revert bytes fixes

+++ beacon-chain/db/kv/blocks.go
@@ -343,9 +343,9 @@ func (k *Store) HighestSlotBlocksBelow(ctx context.Context, slot uint64) ([]*eth
 		sBkt := tx.Bucket(slotsHasObjectBucket)
 		savedSlots := sBkt.Get(savedBlockSlotsKey)
 		if len(savedSlots) == 0 {
-			savedSlots = bytesutil.MakeEmptyBitlists(slot)
+			savedSlots = bytesutil.MakeEmptyBitlists(int(slot))
 		}
-		highestIndex, err := bytesutil.HighestBitIndexAt(savedSlots, slot)
+		highestIndex, err := bytesutil.HighestBitIndexAt(savedSlots, int(slot))
 		if err != nil {
 			return err
 		}
@@ -361,11 +361,11 @@ func (k *Store) HighestSlotBlocksBelow(ctx context.Context, slot uint64) ([]*eth
 
 // blocksAtSlotBitfieldIndex retrieves the blocks in DB given the input index. The index represents
 // the position of the slot bitfield the saved block maps to.
-func (k *Store) blocksAtSlotBitfieldIndex(ctx context.Context, tx *bolt.Tx, index uint64) ([]*ethpb.SignedBeaconBlock, error) {
+func (k *Store) blocksAtSlotBitfieldIndex(ctx context.Context, tx *bolt.Tx, index int) ([]*ethpb.SignedBeaconBlock, error) {
 	ctx, span := trace.StartSpan(ctx, "BeaconDB.blocksAtSlotBitfieldIndex")
 	defer span.End()
 
-	highestSlot := index - 1
+	highestSlot := uint64(index - 1)
 	highestSlot = uint64(math.Max(0, float64(highestSlot)))
 
 	if highestSlot == 0 {
@@ -414,7 +414,7 @@ func (k *Store) setBlockSlotBitField(ctx context.Context, tx *bolt.Tx, slot uint
 	tmp := make([]byte, len(slotBitfields))
 	copy(tmp, slotBitfields)
 
-	slotBitfields = bytesutil.SetBit(tmp, slot)
+	slotBitfields = bytesutil.SetBit(tmp, int(slot))
 	return bucket.Put(savedBlockSlotsKey, slotBitfields)
 }
 
@@ -435,7 +435,7 @@ func (k *Store) clearBlockSlotBitField(ctx context.Context, tx *bolt.Tx, slot ui
 	tmp := make([]byte, len(slotBitfields))
 	copy(tmp, slotBitfields)
 
-	slotBitfields = bytesutil.ClearBit(tmp, slot)
+	slotBitfields = bytesutil.ClearBit(tmp, int(slot))
 	return bucket.Put(savedBlockSlotsKey, slotBitfields)
 }
